Fix:Removed-and-creation-of-new-auth-service
