+++ SchoolManagerSystem.Service/Authentications/Implementations/AuthServices.cs
@@ -2,20 +2,26 @@
 using SchoolManagerSystem.Common.DTOs;
 using SchoolManagerSystem.Model.Entities;
 using SchoolManagerSystem.Service.Authentications.Interfaces;
+using SchoolManagerSystem.Service.Token.Interfaces;
 using System;
 using System.Threading.Tasks;
 
 namespace SchoolManagerSystem.Service.Authentications.Implementations
 {
-	public class RegisterService : IRegisterService
+	public class AuthServices : IAuthServices
 	{
 		private readonly UserManager<ApplicationUser> _userManager;
 		private readonly RoleManager<IdentityRole> _roleManager;
-		public RegisterService(UserManager<ApplicationUser> userManager, RoleManager<IdentityRole> roleManager)
+		private readonly IToken _token;
+
+		public AuthServices(UserManager<ApplicationUser> userManager, RoleManager<IdentityRole> roleManager, IToken token)
 		{
 			_userManager = userManager;
 			_roleManager = roleManager;
+			_token = token;
 		}
+
+
 		public async Task<UserRegistrationResponse> Register(UserRegistrationRequest registerRequest, string isRole)
 		{
 			var user = new ApplicationUser
@@ -63,5 +69,34 @@ namespace SchoolManagerSystem.Service.Authentications.Implementations
 
 			return response;
 		}
+
+
+		public async Task<LoginResponse> Login(LoginRequest loginRequest)
+		{
+			var user = await _userManager.FindByEmailAsync(loginRequest.Email);
+			if (user == null)
+			{
+				throw new ArgumentNullException("Email Provided is Invalid");
+			}
+			var isUserFound = await _userManager.CheckPasswordAsync(user, loginRequest.Password);
+			if (!isUserFound)
+			{
+				throw new ArgumentNullException("User does not Exist");
+			}
+			var token = await _token.GetToken(user);
+
+			var response = new LoginResponse
+			{
+				Id = user.Id,
+				FirstName = user.FirstName,
+				LastName = user.LastName,
+				Email = user.Email,
+				UserName = user.UserName,
+				Token = token
+			};
+
+			return response;
+		}
+
 	}
 }
